# apps/web/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Medicly API",
    "version": "1.0.0",
    "description": "Medicly MVP HTTP API. Cookie session: medicly_sid."
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "medicly_sid"
      }
    },
...
  },
  "paths": {
...
    "/v1/appointments": {
      "post": {
        "operationId": "postBookAppointment",
        "tags": [
          "appointments"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "doctorId",
                  "startAt",
                  "format"
                ],
                "properties": {
                  "doctorId": {
                    "type": "string"
                  },
                  "startAt": {
                    "type": "string"
                  },
                  "format": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "offline"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "online"
                        ]
                      }
                    ]
                  },
                  "reason": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "appointment"
                  ],
                  "properties": {
                    "appointment": {
                      "type": "object",
                      "required": [
                        "id",
                        "doctorId",
                        "patientId",
                        "startAt",
                        "endAt",
                        "format",
                        "status",
                        "reason",
                        "visitDurationMinutes",
                        "proposedStartAt",
                        "cancelledBy"
                      ],
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "doctorId": {
                          "type": "string"
                        },
                        "patientId": {
                          "type": "string"
                        },
                        "startAt": {
                          "type": "string"
                        },
                        "endAt": {
                          "type": "string"
                        },
                        "format": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": [
                                "offline"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "online"
                              ]
                            }
                          ]
                        },
                        "status": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": [
                                "Upcoming"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "Reschedule Pending"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "Completed"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "Cancelled"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "Rescheduled"
                              ]
                            }
                          ]
                        },
                        "reason": {
                          "type": "string",
                          "nullable": true
                        },
                        "visitDurationMinutes": {
                          "type": "number"
                        },
                        "proposedStartAt": {
                          "type": "string",
                          "nullable": true
                        },
                        "cancelledBy": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": [
                                "patient"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "doctor"
                              ]
                            }
                          ],
                          "nullable": true
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/appointments/{id}/reschedule": {
      "post": {
        "operationId": "postRescheduleAppointment",
        "tags": [
          "appointments"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "newStartAt",
                  "format"
                ],
                "properties": {
                  "newStartAt": {
                    "type": "string"
                  },
                  "format": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "offline"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "online"
                        ]
                      }
                    ]
                  },
                  "reason": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "oldAppointment",
                    "newAppointment"
                  ],
                  "properties": {
                    "oldAppointment": {
                      "type": "object",
                      "required": [
                        "id",
                        "doctorId",
                        "patientId",
                        "startAt",
                        "endAt",
                        "format",
                        "status",
                        "reason",
                        "visitDurationMinutes",
                        "proposedStartAt",
                        "cancelledBy"
                      ],
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "doctorId": {
                          "type": "string"
                        },
                        "patientId": {
                          "type": "string"
                        },
                        "startAt": {
                          "type": "string"
                        },
                        "endAt": {
                          "type": "string"
                        },
                        "format": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": [
                                "offline"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "online"
                              ]
                            }
                          ]
                        },
                        "status": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": [
                                "Upcoming"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "Reschedule Pending"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "Completed"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "Cancelled"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "Rescheduled"
                              ]
                            }
                          ]
                        },
                        "reason": {
                          "type": "string",
                          "nullable": true
                        },
                        "visitDurationMinutes": {
                          "type": "number"
                        },
                        "proposedStartAt": {
                          "type": "string",
                          "nullable": true
                        },
                        "cancelledBy": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": [
                                "patient"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "doctor"
                              ]
                            }
                          ],
                          "nullable": true
                        }
                      }
                    },
                    "newAppointment": {
                      "type": "object",
                      "required": [
                        "id",
                        "doctorId",
                        "patientId",
                        "startAt",
                        "endAt",
                        "format",
                        "status",
                        "reason",
                        "visitDurationMinutes",
                        "proposedStartAt",
                        "cancelledBy"
                      ],
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "doctorId": {
                          "type": "string"
                        },
                        "patientId": {
                          "type": "string"
                        },
                        "startAt": {
                          "type": "string"
                        },
                        "endAt": {
                          "type": "string"
                        },
                        "format": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": [
                                "offline"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "online"
                              ]
                            }
                          ]
                        },
                        "status": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": [
                                "Upcoming"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "Reschedule Pending"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "Completed"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "Cancelled"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "Rescheduled"
                              ]
                            }
                          ]
                        },
                        "reason": {
                          "type": "string",
                          "nullable": true
                        },
                        "visitDurationMinutes": {
                          "type": "number"
                        },
                        "proposedStartAt": {
                          "type": "string",
                          "nullable": true
                        },
                        "cancelledBy": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": [
                                "patient"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "doctor"
                              ]
                            }
                          ],
                          "nullable": true
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/appointments/{id}/cancel": {
      "post": {
        "operationId": "postCancelAppointment",
        "tags": [
          "appointments"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "appointment"
                  ],
                  "properties": {
                    "appointment": {
                      "type": "object",
                      "required": [
                        "id",
                        "doctorId",
                        "patientId",
                        "startAt",
                        "endAt",
                        "format",
                        "status",
                        "reason",
                        "visitDurationMinutes",
                        "proposedStartAt",
                        "cancelledBy"
                      ],
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "doctorId": {
                          "type": "string"
                        },
                        "patientId": {
                          "type": "string"
                        },
                        "startAt": {
                          "type": "string"
                        },
                        "endAt": {
                          "type": "string"
                        },
                        "format": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": [
                                "offline"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "online"
                              ]
                            }
                          ]
                        },
                        "status": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": [
                                "Upcoming"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "Reschedule Pending"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "Completed"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "Cancelled"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "Rescheduled"
                              ]
                            }
                          ]
                        },
                        "reason": {
                          "type": "string",
                          "nullable": true
                        },
                        "visitDurationMinutes": {
                          "type": "number"
                        },
                        "proposedStartAt": {
                          "type": "string",
                          "nullable": true
                        },
                        "cancelledBy": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": [
                                "patient"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "doctor"
                              ]
                            }
                          ],
                          "nullable": true
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/appointments/{id}/complete": {
      "post": {
        "operationId": "postCompleteAppointment",
        "tags": [
          "appointments"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "appointment"
                  ],
                  "properties": {
                    "appointment": {
                      "type": "object",
                      "required": [
                        "id",
                        "doctorId",
                        "patientId",
                        "startAt",
                        "endAt",
                        "format",
                        "status",
                        "reason",
                        "visitDurationMinutes",
                        "proposedStartAt",
                        "cancelledBy"
                      ],
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "doctorId": {
                          "type": "string"
                        },
                        "patientId": {
                          "type": "string"
                        },
                        "startAt": {
                          "type": "string"
                        },
                        "endAt": {
                          "type": "string"
                        },
                        "format": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": [
                                "offline"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "online"
                              ]
                            }
                          ]
                        },
                        "status": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": [
                                "Upcoming"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "Reschedule Pending"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "Completed"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "Cancelled"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "Rescheduled"
                              ]
                            }
                          ]
                        },
                        "reason": {
                          "type": "string",
                          "nullable": true
                        },
                        "visitDurationMinutes": {
                          "type": "number"
                        },
                        "proposedStartAt": {
                          "type": "string",
                          "nullable": true
                        },
                        "cancelledBy": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": [
                                "patient"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "doctor"
                              ]
                            }
                          ],
                          "nullable": true
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/api",
      "description": "Vite proxy / same-origin /api"
    }
  ],
...
}